+++ src/nexgen/nxs_write/nxmx_writer.py
@@ -17,7 +17,7 @@ from ..nxs_utils.detector import Detector
 from ..nxs_utils.goniometer import Goniometer
 from ..nxs_utils.sample import Sample
 from ..nxs_utils.source import Attenuator, Beam, Source
-from ..tools.vds_tools import (
+from ..tools.vds_w_tools import (
     clean_unused_links,
     image_vds_writer,
     jungfrau_vds_writer,
